requirements.txt
asgiref==3.7.2
Django==4.2.6
pip==22.0.4
pypng==0.0.21
PyQRCode==1.2.1
setuptools==58.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3